Shaders: ushort to uint

+++ src/shader/cylinder.frag
@@ -8,7 +8,7 @@ uniform float u_cylRad;
 smooth in vec3		   viewImpPos;
 flat in vec3		   viewCylVert[ 2 ];
 flat in vec3		   colors[ 2 ];
-flat in unsigned short vertexSel[ 2 ];
+flat in uint vertexSel[ 2 ];
 
 // 3 16 bits for position.
 // 3 16 bits for normal.
